Remove duplicated code
remove some duplicated code in GMHooks
lots of lines of codes were copied just to check a string check from one
variable to the other.
Simply used one variable and set it to the correct value.
In two cases the copiued code had some extra behavior, i'm not 100% sure
whether that is deliberate or accidental, probably an accident to be
fixed later